Restoring UserDTO to previous location, renaming domain package to dto
