+++ src/main/java/com/agilethought/atsceapi/mapper/UserMapping.java
@@ -2,8 +2,8 @@ package com.agilethought.atsceapi.mapper;
 
 import org.springframework.stereotype.Component;
 
-import com.agilethought.atsceapi.domain.NewUserRequest;
-import com.agilethought.atsceapi.domain.NewUserResponse;
+import com.agilethought.atsceapi.dto.NewUserRequest;
+import com.agilethought.atsceapi.dto.NewUserResponse;
 import com.agilethought.atsceapi.dto.UserDTO;
 import com.agilethought.atsceapi.model.User;
