+++ src/main/java/com/agilethought/atsceapi/service/UserService.java
@@ -2,9 +2,9 @@ package com.agilethought.atsceapi.service;
 
 import java.util.List;
 
-import com.agilethought.atsceapi.domain.LoginData;
-import com.agilethought.atsceapi.domain.NewUserRequest;
-import com.agilethought.atsceapi.domain.NewUserResponse;
+import com.agilethought.atsceapi.dto.LoginData;
+import com.agilethought.atsceapi.dto.NewUserRequest;
+import com.agilethought.atsceapi.dto.NewUserResponse;
 import com.agilethought.atsceapi.dto.UserDTO;
 
 public interface UserService {
